Paralelizaciones y cambios del calculations

+++ calculations.hpp
@@ -11,13 +11,13 @@ extern ofstream accelarationFile;
 extern ofstream velocitiesFile;
 extern ofstream positionsFile;
 
-vec calcForces(double num_objetos, object objetos, vec forces, int iteration);
+vec calcForces(int num_objetos, object objetos, vec forces, int iteration);
 
-vec calcAccelerations(double num_objetos,object objetos, vec forces, vec accelerations, int iteration);
+vec calcAccelerations(int num_objetos,object objetos, vec forces, vec accelerations, int iteration);
 
-void calcVelocities (double num_objetos, object objetos, vec accelerations, int iteration, double incr_tiempo);
+void calcVelocities (int num_objetos, object objetos, vec accelerations, int iteration, double incr_tiempo);
 
-void calcPositions (double num_objetos, object objetos, int iteration, double incr_tiempo, double lado);
+void calcPositions (int num_objetos, object objetos, int iteration, double incr_tiempo, double lado);
 
 
 #endif //PROYECTO_SOA_CALCULATIONS_HPP
